+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix name and namespace regex used in create project from example command
+- Fix `compote autocomplete` incompatible with click 7.x issue
+- Fix `compote autocomplete` failed when .zshrc has no `compinit` set
 
 ## [1.2.1] 2022-12-12